+++ UsefulWebApps/Models/ViewModels/MyRecipes/AddRecipeIngredientToGroceryVM.cs
@@ -4,7 +4,7 @@ namespace UsefulWebApps.Models.ViewModels.MyRecipes
 {
     public class AddRecipeIngredientToGroceryVM
     {
-        public int RecipeId { get; set; }
+        public long RecipeId { get; set; }
 
         [StringLength(100, MinimumLength = 3, ErrorMessage = "Please enter at least 3 characters.")]
         [Required(ErrorMessage = "Grocery item is required.")]
