+++ UsefulWebApps/Repository/IRepository/IRepository.cs
@@ -20,11 +20,11 @@ namespace UsefulWebApps.Repository.IRepository
         /// <returns>A List of all the returned rows</returns>
         Task<IEnumerable<T>> GetAllWhere(string column, string value);
         Task<T> GetByUserId(string? userId);
-        Task<T> GetById(int? id);
+        Task<T> GetById(long? id);
         Task<T> GetRandomRow();
         Task<bool> Add(T entity);
         Task<bool> Update(T entity);
-        Task<bool> Delete(int? id);
+        Task<bool> Delete(long? id);
         //txn method
         Task<bool> DeleteAll();
         /// <summary>
